requirements.txt
Flask==3.1.1
python-dotenv==1.1.1
Requests==2.32.4
slack_sdk==3.35.0
slackeventsapi==3.0.3
urllib3==2.5.0
